+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Better handling of cases when text and image embeddings are mixed in VectorStore
+- Adding optional model_type in WithConstructionConfig and the validation of the model in from_config method (#237)
 
 ## 0.10.0 (2025-03-17)